% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbvnorm.R
\name{rbvnorm}
\alias{rbvnorm}
\title{rbvnorm}
\usage{
rbvnorm(iter, mu, sigma)
}
\arguments{
\item{iter}{number iterations to run in the sample}

\item{mu}{mean vector of distribution}

\item{sigma}{var/cov matrix of distribution}
}
\value{
invisible list of dataframe, iter, mu, sigma
}
\description{
rbvnorm
}
\examples{
\dontrun{rbvnorm(iter, mu, sigma)}

}
